Move requirement lowering down to uv-distribution
