+++ crates/uv-dev/Cargo.toml
@@ -25,11 +25,11 @@ uv-build = { workspace = true }
 uv-cache = { workspace = true, features = ["clap"] }
 uv-client = { workspace = true }
 uv-configuration = { workspace = true }
+uv-distribution = { workspace = true, features = ["schemars"] }
 uv-dispatch = { workspace = true }
 uv-fs = { workspace = true }
 uv-installer = { workspace = true }
 uv-interpreter = { workspace = true }
-uv-requirements = { workspace = true, features = ["schemars"] }
 uv-resolver = { workspace = true }
 uv-types = { workspace = true }
 uv-workspace = { workspace = true, features = ["schemars"] }
