+++ crates/uv-distribution/Cargo.toml
@@ -23,27 +23,40 @@ platform-tags = { workspace = true }
 pypi-types = { workspace = true }
 uv-cache = { workspace = true }
 uv-client = { workspace = true }
+uv-configuration = { workspace = true }
 uv-extract = { workspace = true }
 uv-fs = { workspace = true, features = ["tokio"] }
 uv-git = { workspace = true }
 uv-normalize = { workspace = true }
 uv-types = { workspace = true }
-uv-configuration = { workspace = true }
+uv-warnings = { workspace = true }
 
 anyhow = { workspace = true }
 fs-err = { workspace = true }
 futures = { workspace = true }
+glob = { workspace = true }
+indexmap =  { workspace = true }
 nanoid = { workspace = true }
 once_cell = { workspace = true }
+path-absolutize = { workspace = true }
 reqwest = { workspace = true }
 reqwest-middleware = { workspace = true }
 rmp-serde = { workspace = true }
 rustc-hash = { workspace = true }
+schemars = { workspace = true, optional = true }
 serde = { workspace = true, features = ["derive"] }
 tempfile = { workspace = true }
 thiserror = { workspace = true }
 tokio = { workspace = true }
 tokio-util = { workspace = true, features = ["compat"] }
+toml =  { workspace = true }
 tracing = { workspace = true }
 url = { workspace = true }
 zip = { workspace = true }
+
+[dev-dependencies]
+insta = { version = "1.39.0", features = ["filters", "json", "redactions"] }
+regex = { workspace = true }
+
+[features]
+schemars = ["dep:schemars"]
