+++ crates/uv-distribution/src/lib.rs
@@ -5,7 +5,10 @@ pub use error::Error;
 pub use git::{git_url_to_precise, is_same_reference};
 pub use index::{BuiltWheelIndex, RegistryWheelIndex};
 use pypi_types::{HashDigest, Metadata23};
+pub use pyproject::*;
 pub use reporter::Reporter;
+pub use requirement_lowering::{lower_requirement, lower_requirements, LoweringError};
+pub use workspace::{ProjectWorkspace, Workspace, WorkspaceError, WorkspaceMember};
 
 mod archive;
 mod distribution_database;
@@ -14,8 +17,11 @@ mod error;
 mod git;
 mod index;
 mod locks;
+pub mod pyproject;
 mod reporter;
+mod requirement_lowering;
 mod source;
+mod workspace;
 
 /// The metadata associated with an archive.
 #[derive(Debug, Clone)]
