+++ crates/uv-requirements/Cargo.toml
@@ -13,7 +13,6 @@ license.workspace = true
 cache-key = { workspace = true }
 distribution-filename = { workspace = true }
 distribution-types = { workspace = true }
-pep440_rs = { workspace = true }
 pep508_rs = { workspace = true }
 pypi-types = { workspace = true }
 requirements-txt = { workspace = true, features = ["http"] }
@@ -34,26 +33,19 @@ console = { workspace = true }
 ctrlc = { workspace = true }
 fs-err = { workspace = true, features = ["tokio"] }
 futures = { workspace = true }
-glob = { workspace = true }
 indexmap = { workspace = true }
 path-absolutize = { workspace = true }
 rustc-hash = { workspace = true }
 same-file = { workspace = true }
-schemars = { workspace = true, optional = true }
 serde = { workspace = true }
 thiserror = { workspace = true }
 toml = { workspace = true }
 tracing = { workspace = true }
 url = { workspace = true }
 
-[features]
-schemars = ["dep:schemars"]
-
 [dev-dependencies]
 indoc = "2.0.5"
 insta = { version = "1.38.0", features = ["filters", "redactions", "json"] }
-regex = { workspace = true }
-tokio = { workspace = true }
 
 [lints]
 workspace = true
