+++ crates/uv-requirements/src/lib.rs
@@ -3,15 +3,11 @@ pub use crate::source_tree::*;
 pub use crate::sources::*;
 pub use crate::specification::*;
 pub use crate::unnamed::*;
-pub use crate::workspace::*;
 
 mod confirm;
 mod lookahead;
-pub mod pyproject;
-mod requirement_lowering;
 mod source_tree;
 mod sources;
 mod specification;
 mod unnamed;
 pub mod upgrade;
-mod workspace;
